split reactor into focused files, extract DRY helpers
